Add onlyVisible option to ignore hidden elements
In the default configuration even hidden nodes use space. This might
result in an unintuitive layout. Space is reserved for those
nodes, but no content is displayed.

According to
 http://js.cytoscape.org/#style/visibility
elements with the property display:none do not use space and
elements should not use space and not be connected.

However, the cytoscape API

provides a 'hide' method that sets display to 'none' and a
'hidden' function that is just the oposit of the 'visible' function.

Thus a new option 'onlyVisible' was added that the user can define,
if invisible elements should use space.