Update to @dagrejs/dagre and fix packing
So this, commit address a couple things to make it all work:

1. `webpack parses` `.babelrc` and then doesn't use any of it AFAIK so:
  move `.babelrc` config to `webpack.config.js`.

This is also good because we will have two babel steps:

`ESM --> require`, which is NOT a full transpilation, just changing import
method, and then `require --> web-ready javascript`, which is bundling +
es5 or whatever.

2. I upgraded dagre to the better but still unmaintained @dargejs/dagre,
   mixed feelings on that. It works fine. So that's package.json changes
   + src/layout require() changes.

However, unpacked versions that have to fetch dagre should use the old version,
or I can submit a pull request so @dagrejs/dagre provides a web-ready
version. Makes me wonder if import from browser is going to work.